requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
djangorestframework==3.11.0
gitdb==4.0.5
GitPython==3.1.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
psycopg2==2.8.5
python-dateutil==1.5
python-git==2018.2.1
pytz==2020.1
requests==2.23.0
Send2Trash==1.5.0
smmap==3.0.4
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
wincertstore==0.2
